requirements.txt
numpy==1.23.5
transformers==4.18.0
tensorboard==2.17.1
pytorch-lightning==1.9.5
pycocotools==2.0.5
PyYAML==6.0.1
onnx==1.15.0
onnxruntime==1.19.0
onnxruntime-gpu==1.18.1
open3d==0.18.0
opencv-python==4.10.0.84
pillow==10.4.0
plyfile==1.1
torch_tensorrt==2.3.0
tqdm==4.66.4
tensorrt==10.0.1
tensorrt-cu12==10.2.0.post1
tensorrt-cu12-bindings==10.2.0.post1
tensorrt-cu12-libs==10.2.0.post1
torch-cluster==1.6.3
torch-geometric==2.2.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
Cython==3.0.11
pyvips==2024.2